Design choices
I've created two new notebooks, to test configurations of hexagons against the SRTM data, and one to create a heightmap for the Camino de Santiago shell (or any other arbitrary image). From this, I aim to redesign the base for the hexagons to have a shell imprinted on them

+++ HexGrid.py
@@ -167,7 +167,7 @@ def generateHexBase(numHexInRadius, v, HexD, f, baseVal=-20):
     baseFaces = baseFaces + nv # offset the indices by the number of original vertices
     
     #### NEED TO COMBINE baseFaces AND f
-    f = np.vstack((f,edgeFaces,baseFaces)).T
+    f = np.vstack((f,edgeFaces,baseFaces)).T.astype(int)
     
     return v,HexD,f
